requirements.txt
python-socketio[client]
aiohttp
web3